cleanup cli package
