+++ community-adder-template/tests/setup/global.ts
@@ -1,5 +1,5 @@
 import { fileURLToPath } from 'node:url';
-import { setup, type ProjectVariant } from 'sv/test';
+import { setup, type ProjectVariant } from 'sv/testing';
 import type { GlobalSetupContext } from 'vitest/node';
 
 const variants: ProjectVariant[] = ['kit-js', 'kit-ts', 'vite-js', 'vite-ts'];
