+++ packages/cli/commands/add/utils.ts
@@ -1,8 +1,10 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import pc from 'picocolors';
+import { exec } from 'tinyexec';
 import { parseJson } from '@sveltejs/cli-core/parsers';
 import type { Adder, Highlighter, Workspace } from '@sveltejs/cli-core';
+import { resolveCommand, type AgentName } from 'package-manager-detector';
 
 export type Package = {
 	name: string;
@@ -30,8 +32,21 @@ export function getPackageJson(cwd: string): {
 	return { source: packageText, data: data as Package, generateCode };
 }
 
+export async function formatFiles(options: {
+	packageManager: AgentName;
+	cwd: string;
+	paths: string[];
+}): Promise<void> {
+	const args = ['prettier', '--write', '--ignore-unknown', ...options.paths];
+	const cmd = resolveCommand(options.packageManager, 'execute-local', args)!;
+	await exec(cmd.command, cmd.args, {
+		nodeOptions: { cwd: options.cwd, stdio: 'pipe' },
+		throwOnError: true
+	});
+}
+
 export function readFile(cwd: string, filePath: string): string {
-	const fullFilePath = getFilePath(cwd, filePath);
+	const fullFilePath = path.resolve(cwd, filePath);
 
 	if (!fileExists(cwd, filePath)) {
 		return '';
@@ -76,7 +91,7 @@ function alphabetizeProperties(obj: Record<string, string>) {
 }
 
 export function writeFile(workspace: Workspace<any>, filePath: string, content: string): void {
-	const fullFilePath = getFilePath(workspace.cwd, filePath);
+	const fullFilePath = path.resolve(workspace.cwd, filePath);
 	const fullDirectoryPath = path.dirname(fullFilePath);
 
 	if (content && !content.endsWith('\n')) content += '\n';
@@ -89,12 +104,8 @@ export function writeFile(workspace: Workspace<any>, filePath: string, content: 
 }
 
 export function fileExists(cwd: string, filePath: string): boolean {
-	const fullFilePath = getFilePath(cwd, filePath);
+	const fullFilePath = path.resolve(cwd, filePath);
 	return fs.existsSync(fullFilePath);
-}
-
-export function getFilePath(cwd: string, fileName: string): string {
-	return path.join(cwd, fileName);
 }
 
 export const commonFilePaths = {
