+++ packages/cli/commands/check.ts
@@ -1,10 +1,10 @@
-import { execSync } from 'node:child_process';
 import process from 'node:process';
+import { execSync } from 'node:child_process';
 import pc from 'picocolors';
-import * as resolve from 'empathic/resolve';
 import { Command } from 'commander';
+import * as resolve from 'empathic/resolve';
 import { resolveCommand } from 'package-manager-detector/commands';
-import { getUserAgent } from '../common.ts';
+import { getUserAgent } from '../utils/package-manager.ts';
 
 export const check = new Command('check')
 	.description('a CLI for checking your Svelte code')
