+++ packages/cli/commands/migrate.ts
@@ -2,7 +2,7 @@ import { execSync } from 'node:child_process';
 import process from 'node:process';
 import { Command } from 'commander';
 import { resolveCommand } from 'package-manager-detector';
-import { getUserAgent } from '../common.ts';
+import { getUserAgent } from '../utils/package-manager.ts';
 
 export const migrate = new Command('migrate')
 	.description('a CLI for migrating Svelte(Kit) codebases')
