--- packages/cli/index.ts
@@ -1,3 +1,0 @@
-export { create } from '@sveltejs/create';
-export { installAddon } from './lib/install.ts';
-export type { AddonMap, InstallOptions, OptionMap } from './lib/install.ts';
