+++ packages/cli/lib/index.ts
@@ -1,0 +1,3 @@
+export { create } from '@sveltejs/create';
+export { installAddon } from './install.ts';
+export type { AddonMap, InstallOptions, OptionMap } from './install.ts';
