+++ packages/cli/utils/common.ts
@@ -1,0 +1,78 @@
+import pc from 'picocolors';
+import pkg from '../package.json';
+import * as p from '@sveltejs/clack-prompts';
+import type { Argument, HelpConfiguration, Option } from 'commander';
+
+const NO_PREFIX = '--no-';
+let options: readonly Option[] = [];
+
+function getLongFlag(flags: string) {
+	return flags
+		.split(',')
+		.map((f) => f.trim())
+		.find((f) => f.startsWith('--'));
+}
+
+export const helpConfig: HelpConfiguration = {
+	argumentDescription: formatDescription,
+	optionDescription: formatDescription,
+	visibleOptions(cmd) {
+		// hack so that we can access existing options in `optionTerm`
+		options = cmd.options;
+
+		const visible = cmd.options.filter((o) => !o.hidden);
+		const show: Option[] = [];
+		// hide any `--no-` flag variants if there's an existing flag of a similar name
+		// e.g. `--types` and `--no-types` will combine into a single `--[no-]types` flag
+		for (const option of visible) {
+			const flag = getLongFlag(option.flags);
+			if (flag?.startsWith(NO_PREFIX)) {
+				const stripped = flag.slice(NO_PREFIX.length);
+				const isNoVariant = visible.some((o) => getLongFlag(o.flags)?.startsWith(`--${stripped}`));
+				if (isNoVariant) continue;
+			}
+			show.push(option);
+		}
+		return show;
+	},
+	optionTerm(option) {
+		const longFlag = getLongFlag(option.flags);
+		const flag = longFlag?.split(' ').at(0);
+		if (!flag || !longFlag) return option.flags;
+
+		// check if there's a `--no-{flag}` variant
+		const noVariant = `--no-${flag.slice(2)}`;
+		const hasVariant = options.some((o) => getLongFlag(o.flags) === noVariant);
+		if (hasVariant) {
+			return `--[no-]${longFlag.slice(2)}`;
+		}
+
+		return option.flags;
+	}
+};
+
+function formatDescription(arg: Option | Argument): string {
+	let output = arg.description;
+	if (arg.defaultValue !== undefined && String(arg.defaultValue)) {
+		output += pc.dim(` (default: ${JSON.stringify(arg.defaultValue)})`);
+	}
+	if (arg.argChoices !== undefined && String(arg.argChoices)) {
+		output += pc.dim(` (choices: ${arg.argChoices.join(', ')})`);
+	}
+	return output;
+}
+
+type MaybePromise = () => Promise<void> | void;
+
+export async function runCommand(action: MaybePromise): Promise<void> {
+	try {
+		p.intro(`Welcome to the Svelte CLI! ${pc.gray(`(v${pkg.version})`)}`);
+		await action();
+		p.outro("You're all set!");
+	} catch (e) {
+		p.cancel('Operation failed.');
+		if (e instanceof Error) {
+			console.error(e.stack ?? e);
+		}
+	}
+}
